check for changes in database before making a new sqlite file
git-svn-id: https://halldsvn.jlab.org/repos/trunk/home/gluex/bin@21540 e333a0bb-b310-0410-bd06-bbc7ae8e47d2

+++ rcdb_sqlite_create.sh
@@ -1,15 +1,27 @@
 #!/bin/sh
 source /group/halld/Software/build_scripts/gluex_env_jlab.sh
 sqlite_file=/group/halld/Software/calib/rcdb_sqlite/rcdb_`date +%F`.sqlite
-rm -f $sqlite_file
-$CCDB_HOME/scripts/mysql2sqlite/mysql2sqlite.sh -hhallddb -urcdb rcdb | sqlite3 $sqlite_file | grep -v memory
-size=`stat --format=%s $sqlite_file`
-echo size = $size
-if [ $size -gt 110000000 ]
+minutes_since_change=`mysql -hhallddb -urcdb rcdb \
+    -e "select timestampdiff(MINUTE, created, now()) from logs \
+    order by created desc limit 1;" | grep -v timestampdiff`
+echo info: minutes_since_change = $minutes_since_change
+if [ $minutes_since_change -lt 1440 ]
     then
-    cp -pv $sqlite_file /group/halld/www/halldweb/html/dist/rcdb.sqlite.tmp
-    cd /group/halld/www/halldweb/html/dist
-    mv -v rcdb.sqlite.tmp rcdb.sqlite
+    echo info: recent entries exist, creating $sqlite_file
+    rm -f $sqlite_file
+    $CCDB_HOME/scripts/mysql2sqlite/mysql2sqlite.sh -hhallddb -urcdb rcdb \
+	| sqlite3 $sqlite_file | grep -v memory
+    size=`stat --format=%s $sqlite_file`
+    echo info: size = $size
+    if [ $size -gt 110000000 ]
+        then
+	cp -pv $sqlite_file /group/halld/www/halldweb/html/dist/rcdb.sqlite.tmp
+	cd /group/halld/www/halldweb/html/dist
+	mv -v rcdb.sqlite.tmp rcdb.sqlite
+    else
+	echo error: too small to copy
+    fi
 else
-    echo error: too small to copy
+    echo info: no recent entries, not creating $sqlite_file
 fi
+
